fix(api): retuning new user & night id

+++ api/src/models/night.rs
@@ -84,7 +84,7 @@ pub mod responses {
 	#[derive(serde::Serialize, serde::Deserialize, Default, ToSchema)]
 	pub struct CreateNightResponse {
 		pub msg: String,
-		pub data: Option<usize>,
+		pub data: Option<i32>,
 	}
 
 	#[derive(serde::Serialize, serde::Deserialize, Default, ToSchema)]
